WIP merge LearnOpenGL Point Shadows
Currently, the call to texture() in deferred_shading.fs ShadowCalculation seems to break the shader (resulting in it rendering all black, regardless of whether or not the shadow value is actually used anywhere) unless the shadow height is a specific value. Regardless, no shadows appear to calculating correctly. Further debugging is required.