add separate GameObject static flag (fixes wall-mounted cogs)
the "makeStatic" parameter of GameObject now allows a GameObject to specify that it is static without attempting to make its mesh static. This way static GameObjects can share a mesh with dynamic GameObjects, such as the wall-mounted cogs in the hallway demo. In the future it may be decided to allow a Model to contain both a static and dynamic shape, so that both static and dynamic GameObjects may freely use it.