optimization: batch GameObject rendering per model
GameObjects are now stored in an unordered_map indexed by model name. As a result, we are now able to speed up rendering significantly by sending the textures for each submesh to the GPU just once before rendering that submesh for each GameObject. There is still room for improvement however, as we are still not using instanced rendering. A simple new demo map (field) and GameObject class (FoliageGrass) have been added for batch render performance testing.