expose NR_LIGHTS as a uniform - fixes graphics bugs on scene change
The maximum number of lights is still hard-coded for the time being, but the actual number of lights is now set as a uniform at the start of each render cycle. This fixes "phantom lights" influencing the render despite no longer existing when switching to a scene with fewer active lights than before.