additively blend particles only, fix rendering particle centers with scaling
Particle blending is now performed in two passes; one that renders the particles in black, and another that renders them additively. This allows for additive blending between particles, without blending with the existing pixel data behind the particles. In order for this to work, particles utilizing partial transparency must have their transparency premultiplied.