simplify demo building, don't compile code for unselected demos
The demo that will be compiled is now defined via the COMPILE_DEMO macro in stdafx.h. This means undesired demo files no longer need to be manually excluded from the build to avoid multiple-declaration issues / unnecessary compile time. If no demo building is desired (such as if the engine is being used in another project) simply neglect to define the COMPILE_DEMO macro, or give it an unused value.