'castShadows' mapLoader arg, ignoreModelTexturePaths setting, start modeling warehouse scene
castShadows is now a valid named arg parsed by the mapLoader. ignoreModelTexturePaths setting, when enabled, will cause Models to strip the path provided by material names and simply look for the materials in an equivalently named folder in the texture directory. Additionally, instantiateGameObject and instantiateLight now return a pointer to the newly instantiated object for the sake of convenience.