switch from ASSIMP to OpenFBX for model loading (partially functional)
Due to complications with the way ASSIMP handles fbx files, an attempt is being made to transition to the use of OpenFBX instead. This first commit adds the OpenFBX source (ofbx.h/cpp, miniz.h/c) to the project, and replaces all uses of ASSIMP (map loading, GameObject model loading) with OpenFBX. The transition has been partially successful; the 3d carousel demo can be run, but contains graphical and transform errors that need to be resolved. Additionally, some ASSIMP functions and references still need to be removed, along with an update to the readme and precompiled libraries/dlls download. Bitangents have also been removed from the mesh Vertex struct, as OpenFBX does not appear to load them (and they have been unused up to this point regardless).